# api-rest-template/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "api-rest-template",
    "description": "api-rest-template",
    "termsOfService": "https://opensource.org/licenses/MIT",
    "contact": {
      "name": "Raul Bolivar Navas",
      "email": "rasysbox@hotmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8085",
      "description": "api-rest-template"
    }
  ],
  "tags": [
    {
      "name": "Client",
      "description": "Client operations"
    }
  ],
  "paths": {
    "/api/v1/clients/{clientCode}": {
      "put": {
        "tags": [
          "Client"
        ],
        "summary": "Update client",
        "description": "Update client",
        "operationId": "update",
        "parameters": [
          {
            "name": "clientCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Client updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Client"
        ],
        "summary": "Delete client",
        "description": "Delete client",
        "operationId": "delete",
        "parameters": [
          {
            "name": "clientCode",
            "in": "path",
            "description": "Client client Code",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "CLI-1712345678903"
          }
        ],
        "responses": {
          "204": {
            "description": "Client deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/clients": {
      "get": {
        "tags": [
          "Client"
        ],
        "summary": "Get all clients",
        "description": "Get all clients",
        "operationId": "getAllClients",
        "responses": {
          "200": {
            "description": "List of clients",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDto"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No content",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Client"
        ],
        "summary": "Create client",
        "description": "Create client",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Client created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clients/status/{clientCode}": {
      "patch": {
        "tags": [
          "Client"
        ],
        "summary": "Partial update client",
        "description": "Partial update client",
        "operationId": "partialUpdate",
        "parameters": [
          {
            "name": "clientCode",
            "in": "path",
            "description": "Client ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialClientDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Client updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clients/dni/{dni}": {
      "get": {
        "tags": [
          "Client"
        ],
        "summary": "Get client by DNI",
        "description": "Get client by DNI",
        "operationId": "getByDni",
        "parameters": [
          {
            "name": "dni",
            "in": "path",
            "description": "Client DNI",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": 12345678
          }
        ],
        "responses": {
          "200": {
            "description": "Client",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "404": {
            "description": "Client not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clients/client-code/{clientCode}": {
      "get": {
        "tags": [
          "Client"
        ],
        "summary": "Get client by client Code",
        "description": "Get client by client Code",
        "operationId": "getByClientCode",
        "parameters": [
          {
            "name": "clientCode",
            "in": "path",
            "description": "Client Code",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "CLI-1742128362966"
          }
        ],
        "responses": {
          "200": {
            "description": "Client",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "404": {
            "description": "Client not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClientRequestDto": {
        "type": "object",
        "properties": {
          "dni": {
            "type": "string",
            "description": "Client DNI",
            "example": "12345678"
          },
          "name": {
            "type": "string",
            "description": "Client name",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "description": "Password",
            "example": "123456"
          },
          "gender": {
            "type": "string",
            "description": "Client Gender",
            "example": "M"
          },
          "age": {
            "type": "integer",
            "description": "Client age",
            "format": "int32",
            "example": 25
          },
          "address": {
            "type": "string",
            "description": "Client address",
            "example": "Av. Los Pinos 123"
          },
          "phone": {
            "type": "string",
            "description": "Client phone",
            "example": "987654321"
          },
          "clientType": {
            "type": "string",
            "description": "Client type",
            "example": "NATURAL_PERSON"
          },
          "dniType": {
            "type": "string",
            "description": "Client DNI type",
            "example": "CC"
          },
          "active": {
            "type": "boolean"
          }
        },
        "description": "Client request data transfer object"
      },
      "ClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "dni": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "clientCode": {
            "type": "string"
          },
          "clientType": {
            "type": "string"
          },
          "dniType": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "PartialClientDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          }
        },
        "description": "Partial Client DTO"
      }
    }
  }
}